Fix features
