+++ bin/inx-chronicle/src/collector/solidifier.rs
@@ -19,9 +19,9 @@ use thiserror::Error;
 pub enum SolidifierError {
     #[error(transparent)]
     Doc(#[from] DocError),
-    #[cfg(feature = "stardust")]
-    #[error("the INX requester is missing")]
-    MissingInxRequester,
+    #[cfg(all(feature = "stardust", feature = "inx"))]
+    #[error("the stardust INX requester is missing")]
+    MissingStardustInxRequester,
     #[error(transparent)]
     MongoDb(#[from] mongodb::error::Error),
     #[error(transparent)]
@@ -52,16 +52,16 @@ impl Actor for Solidifier {
     }
 }
 
-#[cfg(feature = "stardust")]
-mod stardust {
+#[cfg(all(feature = "stardust", feature = "inx"))]
+mod stardust_inx {
     use std::str::FromStr;
 
     use chronicle::{db::model::sync::SyncRecord, stardust::MessageId};
 
     use super::*;
     use crate::{
-        collector::stardust::MilestoneState,
-        inx::{InxRequest, InxWorker},
+        collector::stardust_inx::MilestoneState,
+        stardust_inx::{StardustInxRequest, StardustInxWorker},
     };
 
     #[async_trait]
@@ -101,10 +101,14 @@ mod stardust {
                             None => {
                                 // Send the state and everything. If the requester finds the message, it will circle
                                 // back.
-                                cx.addr::<InxWorker>()
+                                cx.addr::<StardustInxWorker>()
                                     .await
-                                    .send(InxRequest::get_metadata(*message_id, cx.handle().clone(), ms_state))
-                                    .map_err(|_| SolidifierError::MissingInxRequester)?;
+                                    .send(StardustInxRequest::get_metadata(
+                                        *message_id,
+                                        cx.handle().clone(),
+                                        ms_state,
+                                    ))
+                                    .map_err(|_| SolidifierError::MissingStardustInxRequester)?;
                                 return Ok(());
                             }
                         }
@@ -113,10 +117,14 @@ mod stardust {
                     None => {
                         // Send the state and everything. If the requester finds the message, it will circle
                         // back.
-                        cx.addr::<InxWorker>()
+                        cx.addr::<StardustInxWorker>()
                             .await
-                            .send(InxRequest::get_message(*message_id, cx.handle().clone(), ms_state))
-                            .map_err(|_| SolidifierError::MissingInxRequester)?;
+                            .send(StardustInxRequest::get_message(
+                                *message_id,
+                                cx.handle().clone(),
+                                ms_state,
+                            ))
+                            .map_err(|_| SolidifierError::MissingStardustInxRequester)?;
                         return Ok(());
                     }
                 }
