+++ bin/inx-chronicle/src/config.rs
@@ -9,8 +9,8 @@ use thiserror::Error;
 
 #[cfg(feature = "api")]
 use crate::api::ApiConfig;
-#[cfg(feature = "inx")]
-use crate::inx::InxConfig;
+#[cfg(all(feature = "stardust", feature = "inx"))]
+use crate::stardust_inx::StardustInxConfig;
 
 #[derive(Error, Debug)]
 pub enum ConfigError {
@@ -24,8 +24,8 @@ pub enum ConfigError {
 #[derive(Default, Debug, PartialEq, Eq, Serialize, Deserialize)]
 pub struct ChronicleConfig {
     pub mongodb: MongoDbConfig,
-    #[cfg(feature = "inx")]
-    pub inx: InxConfig,
+    #[cfg(all(feature = "stardust", feature = "inx"))]
+    pub inx: StardustInxConfig,
     #[cfg(feature = "api")]
     pub api: ApiConfig,
 }
@@ -39,9 +39,9 @@ impl ChronicleConfig {
 
     /// Applies the appropriate command line arguments to the [`ChronicleConfig`].
     pub fn apply_cli_args(&mut self, args: super::CliArgs) {
-        #[cfg(feature = "stardust")]
+        #[cfg(all(feature = "stardust", feature = "inx"))]
         if let Some(inx) = args.inx {
-            self.inx = InxConfig::new(inx);
+            self.inx = StardustInxConfig::new(inx);
         }
         if let Some(connect_url) = args.db {
             self.mongodb = MongoDbConfig::new().with_connect_url(connect_url);
