--- bin/inx-chronicle/src/inx/mod.rs
@@ -1,13 +1,0 @@
-// Copyright 2022 IOTA Stiftung
-// SPDX-License-Identifier: Apache-2.0
-
-mod config;
-mod error;
-mod listener;
-mod worker;
-
-pub use self::{
-    config::InxConfig,
-    error::InxWorkerError,
-    worker::{stardust::InxRequest, InxWorker},
-};
