+++ bin/inx-chronicle/src/main.rs
@@ -9,8 +9,8 @@ mod api;
 mod cli;
 mod collector;
 mod config;
-#[cfg(feature = "inx")]
-mod inx;
+#[cfg(all(feature = "stardust", feature = "inx"))]
+mod stardust_inx;
 
 use std::error::Error;
 
@@ -35,8 +35,8 @@ use thiserror::Error;
 
 #[cfg(feature = "api")]
 use self::api::ApiWorker;
-#[cfg(feature = "inx")]
-use self::inx::{InxWorker, InxWorkerError};
+#[cfg(all(feature = "stardust", feature = "inx"))]
+use self::stardust_inx::{StardustInxWorker, StardustInxWorkerError};
 
 #[derive(Debug, Error)]
 pub enum LauncherError {
@@ -81,8 +81,8 @@ impl Actor for Launcher {
 
         cx.spawn_child(Collector::new(db.clone(), 1)).await;
 
-        #[cfg(feature = "inx")]
-        cx.spawn_child(InxWorker::new(config.inx.clone())).await;
+        #[cfg(all(feature = "stardust", feature = "inx"))]
+        cx.spawn_child(StardustInxWorker::new(config.inx.clone())).await;
 
         #[cfg(feature = "api")]
         cx.spawn_child(ApiWorker::new(db, config.api.clone())).await;
@@ -127,13 +127,13 @@ impl HandleEvent<Report<Collector>> for Launcher {
     }
 }
 
-#[cfg(feature = "inx")]
+#[cfg(all(feature = "stardust", feature = "inx"))]
 #[async_trait]
-impl HandleEvent<Report<InxWorker>> for Launcher {
+impl HandleEvent<Report<StardustInxWorker>> for Launcher {
     async fn handle_event(
         &mut self,
         cx: &mut ActorContext<Self>,
-        event: Report<InxWorker>,
+        event: Report<StardustInxWorker>,
         config: &mut Self::State,
     ) -> Result<(), Self::Error> {
         match &event {
@@ -142,39 +142,42 @@ impl HandleEvent<Report<InxWorker>> for Launcher {
             }
             Report::Error(e) => match &e.error {
                 ActorError::Result(e) => match e {
-                    InxWorkerError::ConnectionError(_) => {
+                    StardustInxWorkerError::ConnectionError(_) => {
                         let wait_interval = config.inx.connection_retry_interval;
                         log::info!("Retrying INX connection in {} seconds.", wait_interval.as_secs_f32());
-                        cx.delay(SpawnActor::new(InxWorker::new(config.inx.clone())), wait_interval)?;
+                        cx.delay(
+                            SpawnActor::new(StardustInxWorker::new(config.inx.clone())),
+                            wait_interval,
+                        )?;
                     }
-                    InxWorkerError::InvalidAddress(_) => {
+                    StardustInxWorkerError::InvalidAddress(_) => {
                         cx.shutdown();
                     }
-                    InxWorkerError::ParsingAddressFailed(_) => {
+                    StardustInxWorkerError::ParsingAddressFailed(_) => {
                         cx.shutdown();
                     }
                     // TODO: This is stupid, but we can't use the ErrorKind enum so :shrug:
-                    InxWorkerError::TransportFailed(e) => match e.to_string().as_ref() {
+                    StardustInxWorkerError::TransportFailed(e) => match e.to_string().as_ref() {
                         "transport error" => {
-                            cx.spawn_child(InxWorker::new(config.inx.clone())).await;
+                            cx.spawn_child(StardustInxWorker::new(config.inx.clone())).await;
                         }
                         _ => {
                             cx.shutdown();
                         }
                     },
-                    InxWorkerError::Read(_) => {
+                    StardustInxWorkerError::Read(_) => {
                         cx.shutdown();
                     }
-                    InxWorkerError::Runtime(_) => {
+                    StardustInxWorkerError::Runtime(_) => {
                         cx.shutdown();
                     }
-                    InxWorkerError::ListenerError(_) => {
+                    StardustInxWorkerError::ListenerError(_) => {
                         cx.shutdown();
                     }
-                    InxWorkerError::MissingCollector => {
-                        cx.delay(SpawnActor::new(InxWorker::new(config.inx.clone())), None)?;
+                    StardustInxWorkerError::MissingCollector => {
+                        cx.delay(SpawnActor::new(StardustInxWorker::new(config.inx.clone())), None)?;
                     }
-                    InxWorkerError::FailedToAnswerRequest => {
+                    StardustInxWorkerError::FailedToAnswerRequest => {
                         cx.shutdown();
                     }
                 },
