+++ bin/inx-chronicle/src/stardust_inx/config.rs
@@ -5,7 +5,7 @@ use std::time::Duration;
 
 use serde::{Deserialize, Serialize};
 
-pub use super::InxWorkerError;
+pub use super::error::InxWorkerError;
 
 /// A builder to establish a connection to INX.
 #[must_use]
