+++ bin/inx-chronicle/src/stardust_inx/mod.rs
@@ -1,0 +1,13 @@
+// Copyright 2022 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+mod config;
+mod error;
+mod listener;
+mod worker;
+
+pub use self::{
+    config::InxConfig as StardustInxConfig,
+    error::InxWorkerError as StardustInxWorkerError,
+    worker::{InxRequest as StardustInxRequest, InxWorker as StardustInxWorker},
+};
